requirements.txt
asgiref==3.8.1
Django==5.1.7
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg-binary==3.2.5
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.0.1
ruff==0.9.10
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
urff==1.0
